+++ src/cljs/snapshot/lumo/repl.cljs
@@ -859,8 +859,9 @@
 
 (defn- make-eval-opts []
   (merge
-    {:ns @current-ns}
-    (select-keys @app-opts [:verbose :static-fns :fn-invoke-direct])))
+    {:ns @current-ns
+     :target :nodejs}
+    (select-keys @app-opts [:verbose :static-fns :fn-invoke-direct :checked-arrays])))
 
 (defn- current-alias-map []
   (let [cur-ns @current-ns]
@@ -868,10 +869,9 @@
       (merge (get-in @st [::ana/namespaces cur-ns :requires])
         (get-in @st [::ana/namespaces cur-ns :require-macros])))))
 
-(defn- reader-eof? [msg]
-  (or
-    (identical? "EOF while reading" msg)
-    (identical? "EOF while reading string" msg)))
+(defn- reader-eof? [e]
+  (let [{:keys [ex-kind]} (ex-data e)]
+    (keyword-identical? ex-kind :eof)))
 
 (defn- read-chars
   [reader]
@@ -968,16 +968,19 @@
   (vreset! current-ns (:ns session-state)))
 
 (def ^{:private true
-       :doc "The default state used to initialize a new REPL session."} default-session-state
-  (atom (capture-session-state)))
+       :doc "The default state used to initialize a new REPL session."}
+  default-session-state
+  (volatile! (capture-session-state)))
 
 (defonce ^{:private true
-           :doc "The state for each session, keyed by session ID."} session-states (atom {}))
+           :doc "The state for each session, keyed by session ID."}
+  session-states
+  (volatile! {}))
 
 (defn- ^:export clear-state-for-session
   "Clears the session state for a completed session."
   [session-id]
-  (swap! session-states dissoc session-id))
+  (vswap! session-states dissoc session-id))
 
 (defn- set-session-state-for-session-id!
   "Sets the session state for a given session."
@@ -987,7 +990,7 @@
 (defn- capture-session-state-for-session-id
   "Captures the session state for a given session."
   [session-id]
-  (swap! session-states assoc session-id (capture-session-state)))
+  (vswap! session-states assoc session-id (capture-session-state)))
 
 (defn- ns-form?
   [form]
@@ -1094,7 +1097,7 @@
                 (handle-error error true)))))))
     (catch :default e
       ;; `;;` and `#_`
-      (when-not (identical? (.-message e) "EOF")
+      (when-not (identical? (.-message e) "Unexpected EOF.")
         (handle-error e true)))
     (finally (capture-session-state-for-session-id session-id)))
   nil)
@@ -1120,13 +1123,8 @@
   (try
     (second (repl-read-string form))
     (catch :default e
-      (let [msg (.-message e)]
-        (cond
-          (identical? "EOF" msg) ""
-          (reader-eof? msg) false
-          :else (do
-                  ;(handle-error e)
-                  ""))))))
+      (when-not (reader-eof? e)
+        ""))))
 
 (defn- ^:export run-main
   [main-ns & args]
@@ -1159,19 +1157,30 @@
   (vreset! current-ns (symbol ns-str)))
 
 (defn- setup-assert! [elide-asserts]
-  (set! *assert* (not elide-asserts)))
+  (set! *assert* (not elide-asserts))
+  (vswap! default-session-state assoc :*assert* *assert*))
 
-(defn- ^:export init [repl? verbose cache-path static-fns fn-invoke-direct elide-asserts]
+(defn- setup-print-namespace-maps! [print-namespace-maps]
+  (set! *print-namespace-maps* print-namespace-maps)
+  (vswap! default-session-state assoc :*print-namespace-maps* *print-namespace-maps*))
+
+(defn- ^:export init
+  [repl? verbose cache-path static-fns fn-invoke-direct elide-asserts checked-arrays]
   (vreset! app-opts {:repl? repl?
                      :verbose verbose
                      :cache-path cache-path
                      :static-fns static-fns
                      :fn-invoke-direct fn-invoke-direct
-                     :elide-asserts elide-asserts})
+                     :elide-asserts elide-asserts
+                     :checked-arrays (keyword checked-arrays)})
   (setup-assert! elide-asserts)
-  (set! *print-namespace-maps* repl?)
+  (setup-print-namespace-maps! repl?)
   (common/load-core-analysis-caches st repl?)
-  (deps/index-js-libs))
+  (deps/index-js-libs)
+  (let [index @deps/js-lib-index]
+    (swap! st assoc :js-dependency-index (into index
+                                           (map (fn [[k v]] [(str k) v]))
+                                           index))))
 
 ;; --------------------
 ;; Introspection